Update uAttributes & refactor the code (#2)
* Send the whole uAttributes instead of splitting it.


* Split lib.rs into different files for better reading.


* Fix the failure of the subscriber.


---------
