+++ src/lib.rs
@@ -11,28 +11,17 @@
 // Contributors:
 //   ZettaScale Zenoh Team, <zenoh@zettascale.tech>
 //
-use async_trait::async_trait;
+pub mod rpc;
+pub mod utransport;
+
 use protobuf::{Enum, Message};
 use std::{
     collections::HashMap,
-    sync::{
-        atomic::{AtomicU64, Ordering},
-        Arc, Mutex,
-    },
-    time::Duration,
+    sync::{atomic::AtomicU64, Arc, Mutex},
 };
 use up_rust::{
-    rpc::{CallOptions, RpcClient, RpcClientResult, RpcMapperError, RpcServer},
-    transport::{datamodel::UTransport, validator::Validators},
-    uprotocol::{
-        Data, UAttributes, UCode, UMessage, UMessageType, UPayload, UPayloadFormat, UPriority,
-        UStatus, UUri, UUID,
-    },
-    uri::{
-        serializer::{MicroUriSerializer, UriSerializer},
-        validator::UriValidator,
-    },
-    uuid::builder::UUIDBuilder,
+    uprotocol::{Data, UAttributes, UCode, UPayload, UPayloadFormat, UPriority, UStatus, UUri},
+    uri::serializer::{MicroUriSerializer, UriSerializer},
 };
 use zenoh::{
     config::Config,
@@ -106,10 +95,10 @@ impl UPClientZenoh {
         UPayloadFormat::from_i32(value)
     }
 
-    // TODO: Transformation between uAttributes and attachment
-    //       Transform all members first and then exclude some we don't need
     fn uattributes_to_attachment(uattributes: &UAttributes) -> anyhow::Result<AttachmentBuilder> {
         let mut attachment = AttachmentBuilder::new();
+        attachment.insert("attr", &uattributes.write_to_bytes()?);
+        /* TODO: We send the whole uattributes directly for the time being and do the benchmark later.
         attachment.insert("id", &uattributes.id.write_to_bytes()?);
         attachment.insert(
             "type_",
@@ -139,11 +128,21 @@ impl UPClientZenoh {
         if let Some(traceparent) = uattributes.traceparent.clone() {
             attachment.insert("traceparent", &traceparent);
         }
+        */
         Ok(attachment)
     }
 
-    // TODO: Same as uattributes_to_attachment
     fn attachment_to_uattributes(attachment: &Attachment) -> anyhow::Result<UAttributes> {
+        let uattributes = UAttributes::parse_from_bytes(
+            attachment
+                .get(&"attr".as_bytes())
+                .ok_or(UStatus::fail_with_code(
+                    UCode::INTERNAL,
+                    "Unable to get uAttributes",
+                ))?
+                .as_slice(),
+        )?;
+        /* TODO: We send the whole uattributes directly for the time being and do the benchmark later.
         let mut uattributes = UAttributes::new();
         if let Some(id) = attachment.get(&"id".as_bytes()) {
             let uuid = UUID::parse_from_bytes(&id)?;
@@ -204,6 +203,7 @@ impl UPClientZenoh {
             let traceparent = traceparent.to_string();
             uattributes.traceparent = Some(traceparent);
         }
+        */
         Ok(uattributes)
     }
 
@@ -310,394 +310,6 @@ impl UPClientZenoh {
             .await
             .map_err(|_| UStatus::fail_with_code(UCode::INTERNAL, "Unable to reply with Zenoh"))?;
 
-        Ok(())
-    }
-}
-
-// TODO: Need to check how to use CallOptions
-#[async_trait]
-impl RpcClient for UPClientZenoh {
-    async fn invoke_method(
-        &self,
-        topic: UUri,
-        payload: UPayload,
-        _options: CallOptions,
-    ) -> RpcClientResult {
-        // Validate UUri
-        UriValidator::validate(&topic)
-            .map_err(|_| RpcMapperError::UnexpectedError(String::from("Wrong UUri")))?;
-
-        // TODO: Comment it and check whether uAttributes is not necessary.
-        //// Validate UAttributes
-        //{
-        //    // TODO: Check why the validator doesn't have Send
-        //    let validator = Validators::Request.validator();
-        //    if let Err(e) = validator.validate(&attributes) {
-        //        return Err(RpcMapperError::UnexpectedError(format!(
-        //            "Wrong UAttributes {e:?}",
-        //        )));
-        //    }
-        //}
-
-        // Get Zenoh key
-        let Ok(zenoh_key) = UPClientZenoh::to_zenoh_key_string(&topic) else {
-            return Err(RpcMapperError::UnexpectedError(String::from(
-                "Unable to transform to Zenoh key",
-            )));
-        };
-
-        // Get the data from UPayload
-        let Some(Data::Value(buf)) = payload.data else {
-            // TODO: Assume we only have Value here, no reference for shared memory
-            return Err(RpcMapperError::InvalidPayload(String::from(
-                "Wrong UPayload",
-            )));
-        };
-
-        // TODO: Check how to generate uAttributes, ex: source, reqid...
-        let mut uattributes = UAttributes::new();
-        uattributes.source = Some(topic.clone()).into();
-        uattributes.sink = Some(topic).into();
-        uattributes.reqid = Some(UUIDBuilder::new().build()).into();
-        let Ok(attachment) = UPClientZenoh::uattributes_to_attachment(&uattributes) else {
-            return Err(RpcMapperError::UnexpectedError(String::from(
-                "Invalid uAttributes",
-            )));
-        };
-
-        let value = Value::new(buf.into()).encoding(Encoding::WithSuffix(
-            KnownEncoding::AppCustom,
-            payload.format.value().to_string().into(),
-        ));
-        // TODO: Query should support .encoding
-        // TODO: Adjust the timeout
-        let getbuilder = self
-            .session
-            .get(&zenoh_key)
-            .with_value(value)
-            .with_attachment(attachment.build())
-            .target(QueryTarget::BestMatching)
-            .timeout(Duration::from_millis(1000));
-
-        // Send the query
-        let Ok(replies) = getbuilder.res().await else {
-            return Err(RpcMapperError::UnexpectedError(String::from(
-                "Error while sending Zenoh query",
-            )));
-        };
-
-        let Ok(reply) = replies.recv_async().await else {
-            return Err(RpcMapperError::UnexpectedError(String::from(
-                "Error while receiving Zenoh reply",
-            )));
-        };
-        match reply.sample {
-            Ok(sample) => {
-                let Some(encoding) = UPClientZenoh::to_upayload_format(&sample.encoding) else {
-                    return Err(RpcMapperError::UnexpectedError(String::from(
-                        "Error while parsing Zenoh encoding",
-                    )));
-                };
-                // TODO: Need to check attributes is correct or not
-                Ok(UMessage {
-                    attributes: Some(uattributes).into(),
-                    payload: Some(UPayload {
-                        length: Some(0),
-                        format: encoding.into(),
-                        data: Some(Data::Value(sample.payload.contiguous().to_vec())),
-                        ..Default::default()
-                    })
-                    .into(),
-                    ..Default::default()
-                })
-            }
-            Err(_) => Err(RpcMapperError::UnexpectedError(String::from(
-                "Error while parsing Zenoh reply",
-            ))),
-        }
-    }
-}
-
-#[async_trait]
-impl RpcServer for UPClientZenoh {
-    async fn register_rpc_listener(
-        &self,
-        method: UUri,
-        listener: Box<dyn Fn(Result<UMessage, UStatus>) + Send + Sync + 'static>,
-    ) -> Result<String, UStatus> {
-        // Do the validation
-        UriValidator::validate(&method)
-            .map_err(|_| UStatus::fail_with_code(UCode::INVALID_ARGUMENT, "Invalid topic"))?;
-
-        // Get Zenoh key
-        let zenoh_key = UPClientZenoh::to_zenoh_key_string(&method)?;
-        // Generate listener string for users to delete
-        let hashmap_key = format!(
-            "{}_{:X}",
-            zenoh_key,
-            self.callback_counter.fetch_add(1, Ordering::SeqCst)
-        );
-
-        let query_map = self.query_map.clone();
-        // Setup callback
-        let callback = move |query: Query| {
-            // Create UAttribute
-            let Some(attachment) = query.attachment() else {
-                listener(Err(UStatus::fail_with_code(
-                    UCode::INTERNAL,
-                    "Unable to get attachment",
-                )));
-                return;
-            };
-            let u_attribute = match UPClientZenoh::attachment_to_uattributes(attachment) {
-                Ok(uattributes) => uattributes,
-                Err(e) => {
-                    log::error!("attachment_to_uattributes error: {:?}", e);
-                    listener(Err(UStatus::fail_with_code(
-                        UCode::INTERNAL,
-                        "Unable to decode attribute",
-                    )));
-                    return;
-                }
-            };
-            // Create UPayload
-            let u_payload = match query.value() {
-                Some(value) => {
-                    let Some(encoding) = UPClientZenoh::to_upayload_format(&value.encoding) else {
-                        listener(Err(UStatus::fail_with_code(
-                            UCode::INTERNAL,
-                            "Unable to get payload encoding",
-                        )));
-                        return;
-                    };
-                    UPayload {
-                        length: Some(0),
-                        format: encoding.into(),
-                        data: Some(Data::Value(value.payload.contiguous().to_vec())),
-                        ..Default::default()
-                    }
-                }
-                None => UPayload {
-                    length: Some(0),
-                    format: UPayloadFormat::UPAYLOAD_FORMAT_UNSPECIFIED.into(),
-                    data: None,
-                    ..Default::default()
-                },
-            };
-            // Create UMessage
-            let msg = UMessage {
-                attributes: Some(u_attribute.clone()).into(),
-                payload: Some(u_payload).into(),
-                ..Default::default()
-            };
-            query_map
-                .lock()
-                .unwrap()
-                .insert(u_attribute.reqid.to_string(), query);
-            listener(Ok(msg));
-        };
-        if let Ok(queryable) = self
-            .session
-            .declare_queryable(&zenoh_key)
-            .callback_mut(callback)
-            .res()
-            .await
-        {
-            self.queryable_map
-                .lock()
-                .unwrap()
-                .insert(hashmap_key.clone(), queryable);
-        } else {
-            return Err(UStatus::fail_with_code(
-                UCode::INTERNAL,
-                "Unable to register callback with Zenoh",
-            ));
-        }
-
-        Ok(hashmap_key)
-    }
-    async fn unregister_rpc_listener(&self, method: UUri, listener: &str) -> Result<(), UStatus> {
-        // Do the validation
-        UriValidator::validate(&method)
-            .map_err(|_| UStatus::fail_with_code(UCode::INVALID_ARGUMENT, "Invalid topic"))?;
-        // TODO: Check whether we still need method or not (Compare method with listener?)
-
-        if self
-            .queryable_map
-            .lock()
-            .unwrap()
-            .remove(listener)
-            .is_none()
-        {
-            return Err(UStatus::fail_with_code(
-                UCode::INVALID_ARGUMENT,
-                "Listener doesn't exist",
-            ));
-        }
-
-        Ok(())
-    }
-}
-
-#[async_trait]
-impl UTransport for UPClientZenoh {
-    async fn send(&self, message: UMessage) -> Result<(), UStatus> {
-        let payload = *message.payload.0.ok_or(UStatus::fail_with_code(
-            UCode::INVALID_ARGUMENT,
-            "Invalid uPayload",
-        ))?;
-        let attributes = *message.attributes.0.ok_or(UStatus::fail_with_code(
-            UCode::INVALID_ARGUMENT,
-            "Invalid uAttributes",
-        ))?;
-        let topic = attributes.clone().sink;
-        // Do the validation
-        UriValidator::validate(&topic)
-            .map_err(|_| UStatus::fail_with_code(UCode::INVALID_ARGUMENT, "Invalid topic"))?;
-
-        // Get Zenoh key
-        let zenoh_key = UPClientZenoh::to_zenoh_key_string(&topic)?;
-
-        // Check the type of UAttributes (Publish / Request / Response)
-        match attributes
-            .type_
-            .enum_value()
-            .map_err(|_| UStatus::fail_with_code(UCode::INTERNAL, "Unable to parse type"))?
-        {
-            UMessageType::UMESSAGE_TYPE_PUBLISH => {
-                Validators::Publish
-                    .validator()
-                    .validate(&attributes)
-                    .map_err(|e| {
-                        UStatus::fail_with_code(
-                            UCode::INVALID_ARGUMENT,
-                            format!("Wrong Response UAttributes {e:?}"),
-                        )
-                    })?;
-                self.send_publish(&zenoh_key, payload, attributes).await
-            }
-            UMessageType::UMESSAGE_TYPE_RESPONSE => {
-                Validators::Response
-                    .validator()
-                    .validate(&attributes)
-                    .map_err(|e| {
-                        UStatus::fail_with_code(
-                            UCode::INVALID_ARGUMENT,
-                            format!("Wrong Response UAttributes {e:?}"),
-                        )
-                    })?;
-                self.send_response(&zenoh_key, payload, attributes).await
-            }
-            _ => Err(UStatus::fail_with_code(
-                UCode::INVALID_ARGUMENT,
-                "Wrong Message type in UAttributes",
-            )),
-        }
-    }
-
-    async fn receive(&self, _topic: UUri) -> Result<UMessage, UStatus> {
-        Err(UStatus::fail_with_code(
-            UCode::UNIMPLEMENTED,
-            "Not implemented",
-        ))
-    }
-
-    async fn register_listener(
-        &self,
-        topic: UUri,
-        listener: Box<dyn Fn(Result<UMessage, UStatus>) + Send + Sync + 'static>,
-    ) -> Result<String, UStatus> {
-        // Do the validation
-        UriValidator::validate(&topic)
-            .map_err(|_| UStatus::fail_with_code(UCode::INVALID_ARGUMENT, "Invalid topic"))?;
-
-        // Get Zenoh key
-        let zenoh_key = UPClientZenoh::to_zenoh_key_string(&topic)?;
-        // Generate listener string for users to delete
-        let hashmap_key = format!(
-            "{}_{:X}",
-            zenoh_key,
-            self.callback_counter.fetch_add(1, Ordering::SeqCst)
-        );
-
-        // Setup callback
-        let callback = move |sample: Sample| {
-            // Create UAttribute
-            let Some(attachment) = sample.attachment() else {
-                listener(Err(UStatus::fail_with_code(
-                    UCode::INTERNAL,
-                    "Unable to get attachment",
-                )));
-                return;
-            };
-            let u_attribute = match UPClientZenoh::attachment_to_uattributes(attachment) {
-                Ok(uattributes) => uattributes,
-                Err(e) => {
-                    log::error!("attachment_to_uattributes error: {:?}", e);
-                    listener(Err(UStatus::fail_with_code(
-                        UCode::INTERNAL,
-                        "Unable to decode attribute",
-                    )));
-                    return;
-                }
-            };
-            // Create UPayload
-            let Some(encoding) = UPClientZenoh::to_upayload_format(&sample.encoding) else {
-                listener(Err(UStatus::fail_with_code(
-                    UCode::INTERNAL,
-                    "Unable to get payload encoding",
-                )));
-                return;
-            };
-            let u_payload = UPayload {
-                length: Some(0),
-                format: encoding.into(),
-                data: Some(Data::Value(sample.payload.contiguous().to_vec())),
-                ..Default::default()
-            };
-            // Create UMessage
-            let msg = UMessage {
-                attributes: Some(u_attribute).into(),
-                payload: Some(u_payload).into(),
-                ..Default::default()
-            };
-            listener(Ok(msg));
-        };
-        if let Ok(subscriber) = self
-            .session
-            .declare_subscriber(&zenoh_key)
-            .callback_mut(callback)
-            .res()
-            .await
-        {
-            self.subscriber_map
-                .lock()
-                .unwrap()
-                .insert(hashmap_key.clone(), subscriber);
-        } else {
-            return Err(UStatus::fail_with_code(
-                UCode::INTERNAL,
-                "Unable to register callback with Zenoh",
-            ));
-        }
-
-        Ok(hashmap_key)
-    }
-
-    async fn unregister_listener(&self, topic: UUri, listener: &str) -> Result<(), UStatus> {
-        // Do the validation
-        UriValidator::validate(&topic)
-            .map_err(|_| UStatus::fail_with_code(UCode::INVALID_ARGUMENT, "Invalid topic"))?;
-        // TODO: Check whether we still need topic or not (Compare topic with listener?)
-
-        if !self.subscriber_map.lock().unwrap().contains_key(listener) {
-            return Err(UStatus::fail_with_code(
-                UCode::INVALID_ARGUMENT,
-                "Listener doesn't exist",
-            ));
-        }
-
-        self.subscriber_map.lock().unwrap().remove(listener);
         Ok(())
     }
 }
