+++ src/utransport.rs
@@ -1,0 +1,185 @@
+//
+// Copyright (c) 2024 ZettaScale Technology
+//
+// This program and the accompanying materials are made available under the
+// terms of the Eclipse Public License 2.0 which is available at
+// http://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
+// which is available at https://www.apache.org/licenses/LICENSE-2.0.
+//
+// SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
+//
+// Contributors:
+//   ZettaScale Zenoh Team, <zenoh@zettascale.tech>
+//
+use crate::UPClientZenoh;
+use async_trait::async_trait;
+use std::sync::atomic::Ordering;
+use up_rust::{
+    transport::{datamodel::UTransport, validator::Validators},
+    uprotocol::{Data, UCode, UMessage, UMessageType, UPayload, UStatus, UUri},
+    uri::validator::UriValidator,
+};
+use zenoh::prelude::{r#async::*, Sample};
+
+#[async_trait]
+impl UTransport for UPClientZenoh {
+    async fn send(&self, message: UMessage) -> Result<(), UStatus> {
+        let payload = *message.payload.0.ok_or(UStatus::fail_with_code(
+            UCode::INVALID_ARGUMENT,
+            "Invalid uPayload",
+        ))?;
+        let attributes = *message.attributes.0.ok_or(UStatus::fail_with_code(
+            UCode::INVALID_ARGUMENT,
+            "Invalid uAttributes",
+        ))?;
+        let topic = attributes.clone().sink;
+        // Do the validation
+        UriValidator::validate(&topic)
+            .map_err(|_| UStatus::fail_with_code(UCode::INVALID_ARGUMENT, "Invalid topic"))?;
+
+        // Get Zenoh key
+        let zenoh_key = UPClientZenoh::to_zenoh_key_string(&topic)?;
+
+        // Check the type of UAttributes (Publish / Request / Response)
+        match attributes
+            .type_
+            .enum_value()
+            .map_err(|_| UStatus::fail_with_code(UCode::INTERNAL, "Unable to parse type"))?
+        {
+            UMessageType::UMESSAGE_TYPE_PUBLISH => {
+                Validators::Publish
+                    .validator()
+                    .validate(&attributes)
+                    .map_err(|e| {
+                        UStatus::fail_with_code(
+                            UCode::INVALID_ARGUMENT,
+                            format!("Wrong Response UAttributes {e:?}"),
+                        )
+                    })?;
+                self.send_publish(&zenoh_key, payload, attributes).await
+            }
+            UMessageType::UMESSAGE_TYPE_RESPONSE => {
+                Validators::Response
+                    .validator()
+                    .validate(&attributes)
+                    .map_err(|e| {
+                        UStatus::fail_with_code(
+                            UCode::INVALID_ARGUMENT,
+                            format!("Wrong Response UAttributes {e:?}"),
+                        )
+                    })?;
+                self.send_response(&zenoh_key, payload, attributes).await
+            }
+            _ => Err(UStatus::fail_with_code(
+                UCode::INVALID_ARGUMENT,
+                "Wrong Message type in UAttributes",
+            )),
+        }
+    }
+
+    async fn receive(&self, _topic: UUri) -> Result<UMessage, UStatus> {
+        Err(UStatus::fail_with_code(
+            UCode::UNIMPLEMENTED,
+            "Not implemented",
+        ))
+    }
+
+    async fn register_listener(
+        &self,
+        topic: UUri,
+        listener: Box<dyn Fn(Result<UMessage, UStatus>) + Send + Sync + 'static>,
+    ) -> Result<String, UStatus> {
+        // Do the validation
+        UriValidator::validate(&topic)
+            .map_err(|_| UStatus::fail_with_code(UCode::INVALID_ARGUMENT, "Invalid topic"))?;
+
+        // Get Zenoh key
+        let zenoh_key = UPClientZenoh::to_zenoh_key_string(&topic)?;
+        // Generate listener string for users to delete
+        let hashmap_key = format!(
+            "{}_{:X}",
+            zenoh_key,
+            self.callback_counter.fetch_add(1, Ordering::SeqCst)
+        );
+
+        // Setup callback
+        let callback = move |sample: Sample| {
+            // Create UAttribute
+            let Some(attachment) = sample.attachment() else {
+                listener(Err(UStatus::fail_with_code(
+                    UCode::INTERNAL,
+                    "Unable to get attachment",
+                )));
+                return;
+            };
+            let u_attribute = match UPClientZenoh::attachment_to_uattributes(attachment) {
+                Ok(uattributes) => uattributes,
+                Err(e) => {
+                    log::error!("attachment_to_uattributes error: {:?}", e);
+                    listener(Err(UStatus::fail_with_code(
+                        UCode::INTERNAL,
+                        "Unable to decode attribute",
+                    )));
+                    return;
+                }
+            };
+            // Create UPayload
+            let Some(encoding) = UPClientZenoh::to_upayload_format(&sample.encoding) else {
+                listener(Err(UStatus::fail_with_code(
+                    UCode::INTERNAL,
+                    "Unable to get payload encoding",
+                )));
+                return;
+            };
+            let u_payload = UPayload {
+                length: Some(0),
+                format: encoding.into(),
+                data: Some(Data::Value(sample.payload.contiguous().to_vec())),
+                ..Default::default()
+            };
+            // Create UMessage
+            let msg = UMessage {
+                attributes: Some(u_attribute).into(),
+                payload: Some(u_payload).into(),
+                ..Default::default()
+            };
+            listener(Ok(msg));
+        };
+        if let Ok(subscriber) = self
+            .session
+            .declare_subscriber(&zenoh_key)
+            .callback_mut(callback)
+            .res()
+            .await
+        {
+            self.subscriber_map
+                .lock()
+                .unwrap()
+                .insert(hashmap_key.clone(), subscriber);
+        } else {
+            return Err(UStatus::fail_with_code(
+                UCode::INTERNAL,
+                "Unable to register callback with Zenoh",
+            ));
+        }
+
+        Ok(hashmap_key)
+    }
+
+    async fn unregister_listener(&self, topic: UUri, listener: &str) -> Result<(), UStatus> {
+        // Do the validation
+        UriValidator::validate(&topic)
+            .map_err(|_| UStatus::fail_with_code(UCode::INVALID_ARGUMENT, "Invalid topic"))?;
+        // TODO: Check whether we still need topic or not (Compare topic with listener?)
+
+        if !self.subscriber_map.lock().unwrap().contains_key(listener) {
+            return Err(UStatus::fail_with_code(
+                UCode::INVALID_ARGUMENT,
+                "Listener doesn't exist",
+            ));
+        }
+
+        self.subscriber_map.lock().unwrap().remove(listener);
+        Ok(())
+    }
+}
